Enhance LDC1614 functionality and testing
- Added checks to prevent concurrent polling jobs in LDC1614::probe().
- Implemented identity verification in LDC1614::resetAndReapply() to ensure correct device identification.
- Introduced new private method _verifyIdentityTracked() for manufacturer and device ID checks.
- Updated polling logic to handle identity verification during reset and reapply processes.
- Enhanced test coverage for probe and recovery functions, including tests for busy states and identity checks.
- Improved HIL runner with new expectation checks for address and channel count validation.
- Added unit tests for HIL runner to ensure proper command classification and handling of no-port scenarios.
- Strengthened agent instructions to prioritize simplicity, robustness, and safety in code development.
- Introduced new prompts for I2C library uniformization and stability fixes, focusing on correctness and system stability.

+++ examples/common/Ldc1614Cli.cpp
@@ -513,7 +513,7 @@ void Cli::printDriverHealth() const {
   const uint32_t now = nowMs();
   const uint32_t totalOk = _device.totalSuccess();
   const uint32_t totalFail = _device.totalFailures();
-  const uint32_t total = totalOk + totalFail;
+  const uint64_t total = static_cast<uint64_t>(totalOk) + totalFail;
   const float successRate = (total > 0U)
                                 ? (100.0f * static_cast<float>(totalOk) /
                                    static_cast<float>(total))
